Move filter methods to traits in filters.rs
