+++ src/filters.rs
@@ -1,82 +1,8 @@
-use crate::types::TemplateString;
-use num_bigint::BigInt;
+use crate::error::PyRenderError;
+use crate::render::{Context, IntoBorrowedContent, IntoOwnedContent, Render};
+use crate::types::Argument;
+use crate::{render::Content, types::TemplateString};
 use pyo3::prelude::*;
-
-struct PartsIterator<'t> {
-    variable: &'t str,
-    start: usize,
-}
-
-impl<'t> Iterator for PartsIterator<'t> {
-    type Item = (&'t str, (usize, usize));
-
-    fn next(&mut self) -> Option<Self::Item> {
-        if self.variable.is_empty() {
-            return None;
-        }
-
-        match self.variable.find('.') {
-            Some(index) => {
-                let part = &self.variable[..index];
-                let at = (self.start, index);
-                self.start += index + 1;
-                self.variable = &self.variable[index + 1..];
-                Some((part, at))
-            }
-            None => {
-                let part = self.variable;
-                self.variable = "";
-                Some((part, (self.start, part.len())))
-            }
-        }
-    }
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct Text {
-    pub at: (usize, usize),
-}
-
-impl Text {
-    pub fn new(at: (usize, usize)) -> Self {
-        Self { at }
-    }
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct Variable {
-    pub at: (usize, usize),
-}
-
-impl<'t> Variable {
-    pub fn new(at: (usize, usize)) -> Self {
-        Self { at }
-    }
-
-    pub fn parts(
-        &self,
-        template: TemplateString<'t>,
-    ) -> impl Iterator<Item = (&'t str, (usize, usize))> {
-        let start = self.at.0;
-        let variable = template.content(self.at);
-        PartsIterator { variable, start }
-    }
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub enum ArgumentType {
-    Variable(Variable),
-    Text(Text),
-    TranslatedText(Text),
-    Int(BigInt),
-    Float(f64),
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub struct Argument {
-    pub at: (usize, usize),
-    pub argument_type: ArgumentType,
-}
 
 #[derive(Debug)]
 pub enum FilterType {
@@ -88,20 +14,158 @@ pub enum FilterType {
     Lower(LowerFilter),
 }
 
+pub trait Applicable<'t, 'py> {
+    fn apply(
+        variable: Option<Content<'t, 'py>>,
+        context: &mut Context,
+    ) -> Result<Option<Content<'t, 'py>>, PyRenderError>;
+}
+
+pub trait ApplicableArg<'t, 'py> {
+    fn apply(
+        variable: Option<Content<'t, 'py>>,
+        arg: &Argument,
+        py: Python<'py>,
+        template: TemplateString<'t>,
+        context: &mut Context,
+    ) -> Result<Option<Content<'t, 'py>>, PyRenderError>;
+}
+
+pub trait ApplicableFilter<'t, 'py> {
+    fn apply(
+        filter: &Py<PyAny>,
+        variable: Option<Content<'t, 'py>>,
+        arg: Option<&Argument>,
+        py: Python<'py>,
+        template: TemplateString<'t>,
+        context: &mut Context,
+    ) -> Result<Option<Content<'t, 'py>>, PyRenderError> {
+        let arg = match arg {
+            Some(arg) => arg.resolve(py, template, context)?,
+            None => None,
+        };
+        let filter = filter.bind(py);
+        let value = match arg {
+            Some(arg) => filter.call1((variable, arg))?,
+            None => filter.call1((variable,))?,
+        };
+        Ok(Some(Content::Py(value)))
+    }
+}
+
 #[derive(Debug)]
 pub struct AddSlashesFilter;
+
+impl<'t, 'py> Applicable<'t, 'py> for AddSlashesFilter {
+    fn apply(
+        variable: Option<Content<'t, 'py>>,
+        context: &mut Context,
+    ) -> Result<Option<Content<'t, 'py>>, PyRenderError> {
+        let content = match variable {
+            Some(content) => content
+                .render(context)?
+                .replace(r"\", r"\\")
+                .replace("\"", "\\\"")
+                .replace("'", r"\'")
+                .into_content(),
+            None => "".into_content(),
+        };
+        Ok(content)
+    }
+}
 
 #[derive(Debug)]
 pub struct AddFilter;
 
+impl<'t, 'py> ApplicableArg<'t, 'py> for AddFilter {
+    fn apply(
+        variable: Option<Content<'t, 'py>>,
+        arg: &Argument,
+        py: Python<'py>,
+        template: TemplateString<'t>,
+        context: &mut Context,
+    ) -> Result<Option<Content<'t, 'py>>, PyRenderError> {
+        let variable = match variable {
+            Some(left) => left,
+            None => return Ok(None),
+        };
+        let right = arg
+            .resolve(py, template, context)?
+            .expect("missing argument in context should already have raised");
+        match (variable.to_bigint(), right.to_bigint()) {
+            (Some(variable), Some(right)) => return Ok(Some(Content::Int(variable + right))),
+            _ => {
+                let variable = variable.to_py(py);
+                let right = right.to_py(py);
+                match variable.add(right) {
+                    Ok(sum) => return Ok(Some(Content::Py(sum))),
+                    Err(_) => return Ok(None),
+                }
+            }
+        }
+    }
+}
+
 #[derive(Debug)]
 pub struct CapfirstFilter;
+
+impl<'t, 'py> Applicable<'t, 'py> for CapfirstFilter {
+    fn apply(
+        variable: Option<Content<'t, 'py>>,
+        context: &mut Context,
+    ) -> Result<Option<Content<'t, 'py>>, PyRenderError> {
+        let content = match variable {
+            Some(content) => {
+                let content_string = content.render(context)?.into_owned();
+                let mut chars = content_string.chars();
+                let first_char = match chars.next() {
+                    Some(c) => c.to_uppercase(),
+                    None => return Ok("".into_content()),
+                };
+                let string: String = first_char.chain(chars).collect();
+                string.into_content()
+            }
+            None => "".into_content(),
+        };
+        Ok(content)
+    }
+}
 
 #[derive(Debug)]
 pub struct DefaultFilter;
 
+impl<'t, 'py> ApplicableArg<'t, 'py> for DefaultFilter {
+    fn apply(
+        variable: Option<Content<'t, 'py>>,
+        arg: &Argument,
+        py: Python<'py>,
+        template: TemplateString<'t>,
+        context: &mut Context,
+    ) -> Result<Option<Content<'t, 'py>>, PyRenderError> {
+        let content = match variable {
+            Some(left) => Some(left),
+            None => arg.resolve(py, template, context)?,
+        };
+        Ok(content)
+    }
+}
+
 #[derive(Debug)]
 pub struct ExternalFilter;
+impl<'t, 'py> ApplicableFilter<'t, 'py> for ExternalFilter {}
 
 #[derive(Debug)]
 pub struct LowerFilter;
+
+impl<'t, 'py> Applicable<'t, 'py> for LowerFilter {
+    fn apply(
+        variable: Option<Content<'t, 'py>>,
+        context: &mut Context,
+    ) -> Result<Option<Content<'t, 'py>>, PyRenderError> {
+        let content = match variable {
+            Some(content) => content.render(context)?.to_lowercase().into_content(),
+            None => "".into_content(),
+        };
+        Ok(content)
+    }
+}
